pose estimate - dense: coding

+++ src/stereo_matching.h
@@ -17,12 +17,12 @@ public:
 	};
 
 	bool init(int method = STEREOMATCH_SSD);
-	u8* stereo_matching(const u32 *image1, const CameraPose &p1, const u32 *image2, const CameraPose &p2, int max_diff = 15, int radius = 2);
-	u8* stereo_matching(const u32 *image1, const u32 *image2, float baseline, int max_diff = 15, int radius = 2);
+	float* stereo_matching(const u32 *image1, const CameraPose &p1, const u32 *image2, const CameraPose &p2, int max_diff = 15, int radius = 2);
+	float* stereo_matching(const u32 *image1, const u32 *image2, float baseline, int max_diff = 15, int radius = 2);
 	
 	void disparity_estimation(const u32 const *image1, const u32 const *image2, float **delta1, float **delta2, const int& radius = 2);
 	void disparity_estimationu8(const u32 const *image1, const u32 const *image2, u8 **delta1, u8 **delta2, const int& radius = 2);
-	u8* lrcheck_and_depth(const float const *delta1, const float const *delta2, float baseline, int max_diff = 15);
+	float* lrcheck_and_depth(const float const *delta1, const float const *delta2, float baseline, int max_diff = 15);
 
 	bool destroy();
